+++ packages/access-definitions/src/Wrapping/WrapperFactories/WrapperObject.php
@@ -126,9 +126,12 @@ class WrapperObject
         if (array_key_exists($propertyName, $readableProperties[0])) {
             $readability = $readableProperties[0][$propertyName];
             $propertyValue = $this->propertyAccessor->getValueByPropertyPath($this->entity, ...$propertyPath);
+            if (!$readability->isValidValue($propertyValue)) {
+                $stringPath = implode('.', $propertyPath);
+                throw new InvalidArgumentException("Value retrieved via property path '$stringPath' is not allowed by readability settings.");
+            }
 
             // if non-relationship, simply use the value read from the target
-            // TODO: validate property type
             return $propertyValue;
         }
 
@@ -337,8 +340,11 @@ class WrapperObject
 
         if (array_key_exists($propertyName, $updatabilities[0])) {
             $updatability = $updatabilities[0][$propertyName];
-            // TODO: add value type validation
-            // TODO: add support for value conditions evaluating non-objects
+
+            // TODO: simplify by adding support for conditions evaluating non-objects
+            if (!$updatability->isValidValue($propertyValue)) {
+                throw new InvalidArgumentException('Value to set into attribute is not allowed by updatability settings.');
+            }
             if (is_object($propertyValue) && !$this->conditionEvaluator->evaluateConditions($propertyValue, $updatability->getValueConditions())) {
                 throw new InvalidArgumentException('Value to set into attribute is not allowed by value conditions.');
             }
@@ -351,7 +357,9 @@ class WrapperObject
             }
         }
 
-        // set via propertyAccessor if there was no custom-write function
+        // Set via propertyAccessor if there was no custom-write function.
+        // We already ensured before this method was called that the property name exists in one
+        // of the updatability-arrays.
         $target = [] === $propertyPath
             ? $this->entity
             : $this->propertyAccessor->getValueByPropertyPath($this->entity, ...$propertyPath);