+++ packages/access-definitions/tests/data/Types/TestAttributeUpdatability.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\data\Types;
+
+use EDT\Wrapping\Properties\AttributeUpdatability;
+
+class TestAttributeUpdatability extends AttributeUpdatability
+{
+    public function isValidValue(mixed $attributeValue): bool
+    {
+        return true;
+    }
+}