+++ packages/jsonapi/src/Properties/AttributeConfig.php
@@ -7,17 +7,17 @@ namespace EDT\JsonApi\Properties;
 use EDT\JsonApi\ResourceTypes\ResourceTypeInterface;
 use EDT\Querying\Contracts\PathsBasedInterface;
 use EDT\Wrapping\Contracts\Types\TransferableTypeInterface;
-use EDT\Wrapping\Contracts\Types\TypeInterface;
 use EDT\Wrapping\Properties\AttributeReadability;
 use EDT\Wrapping\Properties\AttributeUpdatability;
 
 /**
  * @template TCondition of PathsBasedInterface
  * @template TEntity of object
+ * @template TValue
  *
  * @template-extends AbstractConfig<TCondition, TEntity, AttributeReadability<TEntity>, AttributeUpdatability<TCondition, TEntity>>
  */
-class AttributeConfig extends AbstractConfig
+abstract class AttributeConfig extends AbstractConfig
 {
     /**
      * @param ResourceTypeInterface<TCondition, PathsBasedInterface, TEntity> $type
@@ -27,7 +27,7 @@ class AttributeConfig extends AbstractConfig
     ) {}
 
     /**
-     * @param null|callable(TEntity): (simple_primitive|array<int|string, mixed>|null) $customValueFunction
+     * @param null|callable(TEntity): TValue $customValueFunction
      *
      * @return $this
      *
@@ -40,7 +40,7 @@ class AttributeConfig extends AbstractConfig
     ): self {
         $this->assertNullOrImplements(TransferableTypeInterface::class, 'readable');
 
-        $this->readability = new AttributeReadability(
+        $this->readability = $this->createAttributeReadability(
             $defaultField,
             $allowingInconsistencies,
             $customValueFunction
@@ -50,9 +50,20 @@ class AttributeConfig extends AbstractConfig
     }
 
     /**
-     * @param list<TCondition>                                 $entityConditions
-     * @param list<TCondition>                                 $valueConditions
-     * @param null|callable(TEntity, simple_primitive|array<int|string, mixed>|null): void $customWrite
+     * @param null|callable(TEntity): TValue $customValueFunction
+     *
+     * @return AttributeReadability<TEntity>
+     */
+    abstract protected function createAttributeReadability(
+        bool $defaultField,
+        bool $allowingInconsistencies,
+        ?callable $customValueFunction
+    ): AttributeReadability;
+
+    /**
+     * @param list<TCondition> $entityConditions
+     * @param list<TCondition> $valueConditions
+     * @param null|callable(TEntity, TValue): void $customWrite
      *
      * @return $this
      */
@@ -63,7 +74,7 @@ class AttributeConfig extends AbstractConfig
     ): AttributeConfig {
         $this->assertNullOrImplements(TransferableTypeInterface::class, 'readable');
 
-        $this->updatability = new AttributeUpdatability(
+        $this->updatability = new JsonAttributeUpdatability(
             $entityConditions,
             $valueConditions,
             $customWrite