+++ packages/jsonapi/tests/data/ApiTypes/EmptyType.php
@@ -6,13 +6,10 @@ namespace Tests\data\ApiTypes;
 
 use EDT\ConditionFactory\ConditionFactoryInterface;
 use EDT\JsonApi\Properties\TypedPathConfigCollection;
-use EDT\JsonApi\RequestHandling\MessageFormatter;
 use EDT\JsonApi\ResourceTypes\AbstractResourceType;
 use EDT\Querying\Contracts\PathsBasedInterface;
 use EDT\Querying\Contracts\PropertyPathAccessInterface;
 use EDT\Querying\PropertyPaths\PropertyPath;
-use EDT\Wrapping\WrapperFactories\WrapperObjectFactory;
-use Psr\Log\LoggerInterface;
 use Tests\data\EmptyEntity;
 
 class EmptyType extends AbstractResourceType
@@ -23,7 +20,7 @@ class EmptyType extends AbstractResourceType
 
     protected function configureProperties(TypedPathConfigCollection $configCollection): void
     {
-        $configCollection->configureAttribute(
+        $configCollection->configureJsonAttribute(
             new PropertyPath(null, '', PropertyPathAccessInterface::DIRECT, ['id'])
         )->enableReadability(true);
     }